refactor: extract `character_preview_window`
