+++ src/newcharacter.h
@@ -33,55 +33,6 @@ struct points_left {
     std::string to_string();
 };
 
-struct character_preview_window {
-        enum OrientationType : std::uint8_t {
-            TOP_LEFT,
-            TOP_RIGHT,
-            BOTTOM_LEFT,
-            BOTTOM_RIGHT
-        };
-        struct Margin {
-            int left = 0;
-            int right = 0;
-            int top = 0;
-            int bottom = 0;
-        };
-        struct Orientation {
-            OrientationType type = TOP_RIGHT;
-            Margin margin = Margin{};
-        };
-
-        catacurses::window w_preview;
-
-        void init( Character *character );
-        /** Window preparations before displaying. Sets desirable position. Could also be usefull for ui-rescale **/
-        void prepare( int nlines, int ncols, const Orientation *orientation, int hide_below_ncols );
-        void zoom_in();
-        void zoom_out();
-        void toggle_clothes();
-        void display() const;
-        /** Use it as you done with preview **/
-        void clear() const;
-        bool clothes_showing() const;
-
-    private:
-        point pos;
-        int termx_pixels = 0;
-        int termy_pixels = 0;
-        const int MIN_ZOOM = 32;
-        const int MAX_ZOOM = 128;
-        const int DEFAULT_ZOOM = 128;
-        int zoom = DEFAULT_ZOOM;
-        int hide_below_ncols = 0;
-        int ncols_width = 0;
-        int nlines_width = 0;
-        Character *character = nullptr;
-        std::vector<detached_ptr<item>> clothes;
-        bool show_clothes = true;
-
-        point calc_character_pos() const;
-};
-
 namespace newcharacter
 {
 /** Returns the id of a random starting trait that costs >= 0 points */
