+++ src/output.cpp
@@ -64,16 +64,6 @@ int OVERMAP_LEGEND_WIDTH;
 
 scrollingcombattext SCT;
 
-int termx_to_pixel_value()
-{
-    return projected_window_width() / TERMX;
-}
-
-int termy_to_pixel_value()
-{
-    return projected_window_height() / TERMY;
-}
-
 // utf8 version
 std::vector<std::string> foldstring( const std::string &str, int width, const char split )
 {
